separando as classes
